requirements.txt
databricks-connect==15.4.*
databricks-sdk>=0.32.0
pandas>=2.0.0
numpy>=1.26.0
psycopg2-binary>=2.9.9
streamlit>=1.35.0
mlflow>=2.14.0
langgraph>=0.2.0
